Data Wrappers binding
Althought it is currently very limited, data wrappers can now be binded to other data wrappers, which allows them to have their values be synced.
Also added a way for components to handle being moved in memory (which only happens on initialization, that is placing in a container) which allows to update data wrappers binders so the pointers don't point to invalid locations.
This was preferred instead of a handle system as memory location only changes once in a program lifetime.