Exit the event loop when the last window is closed
The event loop will now exit when the last window has been closed. It was also renamed to zgt.runEventLoop() as the event loop is for the whole application and not specific to a window. It can also now be stepped using zgt.stepEventLoop().

This commit also changes backends to use zgt lasting and scratch allocator.