win32: Prepare for drawing with GDI+
The reason one would want to draw with GDI+ instead of GDI is that it's just more recent. It also introduces more recent featueres like alpha blending.

One step above would be to render with OpenGL (not with DirectX, in order to share code with the GLES backend) but this drives up the code complexity and shouldn't be necessary.