Add mobile device 'emulation'
By passing the ZGT_MOBILE_EMULATED environment variable, zgt
will force the window size and source DPI to be equal to the
given phone's one. In the future, this will allow testing
responsive design easily and with slow compile times.
In essence, this is similar to Chrome's and Firefox's
mobile emulation mode.