win32: Fix flickering
The flickering can notably be noticed in the 'colors' example, where clicking the 'Random color' button will cause the button to disappear (you had to move the cursor for it to reappear)

To fix this, WS_EX_COMPOSITED style has been added. While this fixed the disappearance, the button now flickered. Adding WS_EX_LAYERED too seems to have fixed the problem.