Rename DataWrapper to Atom
Atom seems to be the most widely used name for this kind of thing, and besides it a thousand times better than the plain 'DataWrapper' name and shorter.
Atom is also a great communicator of the fact that DataWrapper is meant to be widely used for about any property you have.