simplify component management
Simplify component management by immediately placing them on the heap, instead of the previous
situation where a component was first in the stack, then copied to the heap or not at some mysterious
point in time which depends on the lifecycle, and pointers also had to be updated in a weird, hacky
way.
Making everything allocated on the heap from the get go thus immensely simplifies the behaviour of
components. It also simplifies the handling of Widgets, which are now always pointers and have a much
simpler lifecycle.