build: apply the Console subsystem for Debug builds
If no subsystem is specified, then the resulted binary doesn't
constitute as a valid binary Win32 executable. Hence the need to specify
a subsystem for debug builds too. The `.Console` subsystem was chosen as
it allows for standard output to be logged in a console (which is quite
useful for debug builds)