feat(atom): use monotonic time for animations and add checksum
Monotonic time is now used for animations as it is specified to not have
time leaps (like leap seconds or timezone changes) which is what's
needed for animations.

Checksums have also been added to atoms as this allows to differentiate
a slice's new value from the old one when its content has been changed
in-place.