add safe-guards to event handlers
I noticed it ended up being quite common to accidentally use an invalid
function type, which was allowed because the argument type is `anytype`.
Notably, a common mistake is using `void` as a return type instead of
`!void` or `anyerror!void`, which would cause runtime errors that are
hard to understand and debug.