build(wasm): embed HTML files in the builder
The `prefix` path had stopped working following the latest Zig update,
which means an alternative had to be found that doesn't depend on
`@src().file` (which now only returns paths relative to the module
root). For now, the best solution I've found is to embed the files in
the WebServerStep.