fix: make Atom.hasAnimation side-effect free
Previously, Atom.hasAnimation would call Atom.update, which would in
turn call the Atom's change listeners if the Atom was in an animation.
However, This would cause infinite recursion loop in cases where
Atom.hasAnimation is called inside of or as a result of a change
listener.
This method simply does not need to call change listeners, as neither
its name nor its documentation implies it has side-effects, it doesn't
even lock the Atom as by design, TOCTOU errors can't happen on Atoms.